Refactor configuration to improve environment handling
