+++ Meadowlark-js/package.json
@@ -48,8 +48,10 @@
     "test:lint:ts": "tsc -p . --noEmit",
     "test:unit": "cross-env NODE_OPTIONS=--max-old-space-size=4096 LOG_LEVEL=warn npx jest --testPathIgnorePatterns=\"integration|e2e\"",
     "test:integration": "cross-env NODE_OPTIONS=--max-old-space-size=4096 LOG_LEVEL=warn npx jest integration --runInBand",
-    "test:e2e": "cross-env NODE_OPTIONS=--max-old-space-size=4096 LOG_LEVEL=warn npx jest e2e --config ./jest.e2e-config.js --runInBand",
     "test:unit:coverage": "rimraf coverage/ && npm run test:unit -- --collectCoverage",
+    "test:e2e": "npm run setup:e2e && npm run run:e2e",
+    "setup:e2e": "bash ./tests/e2e/setup/scripts/auth-key-file-setup.sh",
+    "run:e2e": "cross-env NODE_OPTIONS=--max-old-space-size=4096 LOG_LEVEL=warn npx jest e2e --config ./jest.e2e-config.js --runInBand",
     "start:local": "lerna run start:local",
     "docker:lint": "cat ./Dockerfile | docker run --rm -i hadolint/hadolint",
     "docker:install": "npm i lerna@^6.0.2 rimraf@^3.0.2 copyfiles@^2.4.1 typescript@4.8.4 -g && npm ci --only=production",
