--- Meadowlark-js/tests/e2e/.env.example
@@ -1,4 +1,0 @@
-# This is an example file of how to configure environment
-OAUTH_SIGNING_KEY={{OAUTH_SIGNING_KEY}}
-MONGODB_PASS={{MONGODB_PASS}}
-DOCUMENT_STORE_PLUGIN={{DOCUMENT_STORE_PLUGIN}} # defaults to @edfi/meadowlark-mongodb-backend
