+++ Meadowlark-js/tests/e2e/setup/.env.example
@@ -1,0 +1,10 @@
+OAUTH_SIGNING_KEY={{OAUTH_SIGNING_KEY}}
+
+# OPTIONAL SETTINGS
+
+# DOCUMENT_STORE_PLUGIN={{DOCUMENT_STORE_PLUGIN}} # defaults to @edfi/meadowlark-mongodb-backend
+# MONGODB_USER={{MONGODB_USER}}
+# MONGODB_PASS={{MONGODB_PASS}}
+
+# Enable TestContainers Debugging
+# DEBUG=testcontainers*
